network-visibility/requirements.txt
netmiko>=4.1.2
textfsm>=1.1.2
python-dotenv>=1.0.0
networkx>=3.1
matplotlib>=3.7.1